Use .env file for API token
Use a separate .env file for the API token in order to make deployment easier. This way the token can always be present in its own location.
Changes to the file are ignored by .gitignore so that the token doesn't have to be removed from the file.

Also, bored at work.